requirements.txt
PyGithub==2.3.0
python-gitlab==4.8.0
requests==2.32.3
python-Levenshtein==0.25.1
openai==1.51.0
pydantic==2.9.2

